@@ -1,57 +1,43 @@
 import React from "react"
-import Link from "react-router/lib/Link"
+import Link from "gatsby-link"
+import { presets } from "glamor"
 
-import { rhythm } from "utils/typography"
+import typography, { rhythm } from "../../utils/typography"
+import SidebarBody from "../../components/sidebar-body"
 
-const IndexRoute = React.createClass({
+class IndexRoute extends React.Component {
   render () {
-    //console.log('Docs index')
-    //console.log(this.props)
-    const pages = this.props.data.allMarkdownRemark.edges
-    const packagePages = pages.filter(page => page.node.package)
-    const otherPages = pages.filter(page => !page.node.package)
     return (
       <div>
-        <h1>Pages</h1>
-        <h2>General Docs</h2>
-        <ul>
-          {otherPages.map(edge => (
-            <li key={edge.node.slug}>
-              <Link to={`${edge.node.slug}`}>
-                {edge.node.frontmatter.title}
-              </Link>
-            </li>
-            ))}
-        </ul>
-        <h2>Official packages</h2>
-        <ul>
-          {packagePages.map(edge => (
-            <li key={edge.node.slug}>
-              <Link to={`${edge.node.slug}`}>
-                {edge.node.frontmatter.title}
-              </Link>
-            </li>
-            ))}
-        </ul>
+        <h1>Getting started</h1>
+        <p>Gatsby lets you transform plain text into dynamic blogs and websites using
+          the latest web technologies. A React.js static site generator.
+        </p>
+        <p>
+          It supports Markdown, HTML, and React.js pages out of the box. Easy to
+          add support for additional file types.
+        </p>
+        <h2>Install</h2>
+        <p><code>npm install -g gatsby</code></p>
+        <p>TBD best way to get started — official starters still (with official
+          themes underneath?</p>
+        <div
+          css={{
+            display: `block`,
+            [presets.Tablet]: {
+              display: `none`,
+            },
+          }}
+        >
+          <h2>Documentation</h2>
+          <SidebarBody inline={true} />
+        </div>
       </div>
     )
-  },
-})
+  }
+}
 
 export default IndexRoute
 
 export const pageQuery = `
-{
-  allMarkdownRemark {
-    edges {
-      node {
-        slug
-        package
-        frontmatter {
-          title
-        }
-      }
-    }
-  }
-}
 `
