@@ -24,7 +24,7 @@ module.exports = (
 
   // If the users didn't set a default sizes, we'll make one.
   if (!options.sizes) {
-    options.sizes = `(max-width: ${options.maxWidth}px) 100vw, {${options.maxWidth}px}`
+    options.sizes = `(max-width: ${options.maxWidth}px) 100vw, ${options.maxWidth}px`
   }
 
   // Create sizes (in width) for the image. If the max width of the container
